refactor: implement TaskUI module for task UI management
